requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
distro==1.8.0
Django==4.1.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==5.4.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.12
websocket-client==0.59.0
